+++ Tests/ApexyAlamofireTests/AlamofireClientTests.swift
@@ -6,7 +6,7 @@
 //
 
 import Apexy
-import Apexy_Alamofire
+import ApexyAlamofire
 import XCTest
 
 final class AlamofireClientTests: XCTestCase {
